refactor: move sub_tier entity to where it belongs
