+++ src/subscription-tier/subscription-tier.entity.ts
@@ -1,0 +1,14 @@
+import { ColumnType, Insertable, Selectable, Updateable } from "kysely";
+import { Generated } from "../common/kysely-types/generated";
+
+export interface SubscriptionTierTable {
+	id: Generated<string>;
+	tier: string;
+	power: ColumnType<number, number | undefined, number | undefined>;
+	permissions: ColumnType<string[], string[] | undefined, string[] | undefined>;
+	price_rubles: number;
+}
+
+export type SubscriptionTier = Selectable<SubscriptionTierTable>;
+export type NewSubscriptionTier = Insertable<SubscriptionTierTable>;
+export type SubscriptionTierUpdate = Updateable<SubscriptionTierTable>;
